sftp watch remote directory, fixing changes requested
